Show collisions with other balls in trajectory projection, optimise much
I'll next attempt to make it show all balls' trajectories following a hit, I think it can be done on a per-frame basis. Also, it currently assumes the cueball is moving in the projection but other balls are all stationary, it won't account for balls that move into your trajectory because another one hit them in or vice versa, but anything is better than nothing.