Make balls reflect off holes' outer edges instead of falling in when going fast enough
Programming for this was almost finished in previous version, but didn't work (the collision detection loop tells the collider which two balls will collide first in the remainder of the frame simulated, then time is proceeded to the time of their collision, they collide and it finds the next one, continuing until none remain, but the second ball can be substituted for -1 if it collides with a vertical wall, 0 for a horizontal one and (number of balls)+n for nth hole, I forgot to subtract the number of balls in the collider function (which converts it from ball IDs to positions, velocities and masses (radii not necessary, it only needs to know they're touching))).
Also, fix gravity mode not having the grey outside around the circle of play (will eventually be a second, ring-shaped hole, as well as the one they orbit around) at the top and bottom. Also made gravity work with every hole instead of only the origin (no effectual difference yet, may make a binary star system mode or something (I could also generalise holes as a subset of balls in the same list for this, to be neater if I wanted them to be able to move and undergo gravity and such)).
Next, I'll make projection show red circles where balls fall into holes and work in orbital mode (which only showed a single orbit, I may make collisions for all balls show a full orbit from their position and simulate until balls have gone at least an orbit without collision), both of these worked when only the white ball was projected but haven't yet been reimplemented in projecting them all (which does the rendering during the simulating instead of afterwards because lists are 1D).