Add projection for all balls in gravity mode, make outside of disk reflect balls inwards off it
Also re-add the red circle indicators at the ends of trajectories where balls fall into holes.
In gravity mode, the cue ball was initially the only one projected, it would simulate the inverse-square law from the current position and velocity until it had completed a full 360º rotation around the centre. When the first version of the projection taking all balls into account (but assuming all but the cue were stationary) was added, gravity mode's projection would sometimes flicker inwards or outwards depending on whether it would hit the inner or outer side of another ball, then it was disabled when the current system (simulating projected planets in a single function shared with the actual movement) replaced the single-ball one. Now, it projects all balls' trajectories until they've all completed a full rotation or have been deleted.
Parabolic and hyperbolic orbits would never rotate 360º, so the outside of the disk on which it's played configurably either acts as an additional hole or a wall that reflects them inwards (with the same 'crazy mode' in collision equations as the outside of the holes in regular mode).
Also fix problem in previous version's logic that made balls in gravity mode accelerate unreasonably near the hole's centre (gaining velocity and violating Kepler's second law) instead of falling in and being deleted, and optimise much logic.