Make balls' renderings draw lines between all vertices from previous to current frame, like their trajectory projections.
To do so, much had to be rearranged, and the rendering is done during the simulation of a frame instead of afterwards (like in projection), but it's simpler in total, now, I think.
Also, like how you can hold the g key when starting it for the gravity mode, you can now also hold the c key for the Pi calculator mode (not yet finished, will show first ball larger than second and have adjustable precision when interface is added), as seen in https://youtu.be/HEfHFsfGXjs. Only accurate to ~4 digits (calculates it to about 3.14160 due to accumulating floating point inaccuracies in Scratch itself) but a demonstration of the accuracy of the physics in the actual game.
Also, don't try to use the reference frames (with the 2 and 3 keys, keep to 1) in this version, they make balls teleport into each other (which causes many to fall into the black hole in gravity mode), not sure why.