Add n-body gravity mode and make black ball have white outline in gravity mode
In n-body gravity mode (currently inaccessible from frontend), balls are attracted to each other instead of black holes and can cluster (but not accelerate unreasonably, because they can't get to periapses low enough for per-frame granularity to be too slow to violate conservation of energy quickly, only slowly).
The black ball was previously near-invisible against gravity mode's black background, which was quite funny but unintentional.
Also, I found the source of the balls teleporting together when reference frame 2 or 3 was selected, it wasn't the draw vertice function interfering with another use of the dx and dy variables, it was because it had a 'set Target 1 to 1' variable and Target 1 was the iterator for which ball's projected position became ball n's actual one, very frustrating. Now there seem to also be images of balls that don't follow the reference frame angle or position but are rendered on top of everything else, instead, very odd, will fix.
Also considering making holes use the same 'Planet' lists as balls, to simplify the programming and so they can also undergo gravity, perhaps also an indicator of which side their outer wall is on, balls would fall in from relative velocity being too low instead of absolute. Also, perhaps a reference frame mode 4, where the direction of current acceleration is downwards instead of velocity or position being upwards.