Add gravity mode variation with binary black hole system, make holes use planet lists like balls (internal change), simplify some of the collision functions (no functional change, I don't think, only simpler maths).
Hold 'b' while starting for binary mode, like 'g' for regular gravity and 'c' for the Pi calculator. The black holes orbit each other in fixed circular orbits that can't be moved by absorbing balls, imagine balls have 0 mass. The orbits are determined by each of them imagining they're balls orbiting an imaginary hole in the centre with a quarter of the other one's mass (because it's half the distance), they calculate acceleration live each frame with the same programming as the balls.
There will be an option to make gravity work both ways and have proper conservation of momentum (absorbing balls' momentum) when the interface is finished, but the default will keep orbits elliptical in gravity mode and the holes' orbits around each other circular in binary mode.
The list unification is going to make these plans easier, and the gravitate function no longer needs to repeat itself with different list names.
Also, in reference frame mode 1, up is up; in 2, the current player position is up; in 3, the current velocity is; in new mode 4, the current acceleration is down. Excluding collisions, 4 is identical to 3 in regular mode, where the only acceleration is air resistance, and in regular gravity mode, where the black hole is always downwards, it's identical to 2, but in mode 4, with the binary black hole system. 2 behaves as you'd expect, but your motions up and down aren't sinusoidal or oscillating, you move between constantly-changing periapses and apoapses. 3 is like usual, drawing ellipsoid-looking curves, but their ends don't line up. 4 is much wackier, no patterns seem very discernible to human perception, you get closer to being central on the perceived X axis as you near the top if bouncing, and the holes don't appear to orbit each other linearly but also sinusoidally back and forth, it becomes very strange when you move between them. Loops can be manoeuvred into, if you're sufficiently patient.