Make new 'frozen for move' system that pauses live physics simulation after all balls have stopped moving in regular mode (saving computing power) or when the cueball has made a full orbit since the last move in gravity mode (removing its time constraints that wouldn't work well with passing a device or sending gamestate strings), add experimental clicking mode (balls make clicking sound when colliding with volume proportional to magnitude)
The frozen for move system took much longer than expected, every fix for a given problem uncovered further ones (like the black ball's outline in gravity mode only having been programmed in the draw vertice function, not the render one which was previously only used to draw the walls between the corner-rounders (because they need to be connected to each other, not themselves in previous frames, they have infinite mass and don't move), then using the vertice function in render instead of move would make them initially render where they had been before the program had last been reset, somewhere in between those were the balls rendering their first vertice as the centre (so they looked like a spider's legs) and motion trails not decaying (so it was difficult to tell where balls were in their trails, like hard mode for billiards), then all projection in gravity mode entered an infinite loop because when I moved the two lines for determining freezing criteria that referenced the cueball's list number with the 'Cueball Target' variable from Move Planets to Move Projected Planets (so that it could be used to determine how far to project), I didn't also move the line that set the variable).
Scratch limitations prevent multiple concurrent click sounds and make the outside of the table flicker green for a frame whenever it happens (because it seems to override the 'run without screen refresh' option, and the calculation happens in the rendering part of the frame, not the part where it makes the outside black (it can't just avoid making it green, it renders the hole-roundings as circles)), so not enabled, may be a setting but for now set the variable to 1 if you really want it (we think you're going to love it (topical because the WWDC happens today)).