Add RCS propellant particles, exclude them in reference frame mode 4 acceleration calculation (fixes spinning that cancelled out cumulative thrust), fix collisions between massless objects.
(This is primarily an update for the mode with the infinite space and atmosphere accessed with the 'a' key on startup, doesn't affect billiards.)
RCS propellant is fired from the player each frame you have the mouse down. It's generated entirely inside you but touching your edge at a tangent. It has proper rigid-body collisions but very low mass, so if it hits another planet, it will look like it's reflected off. It fires from behind you with conservation of momentum (the average velocity (weighted by mass) of a player and piece of propellant immediately after generation is equal to before, but they deteriorate). Currently, if the distance from you to the surface of the planet you're attempting to liftoff from is lower than the propellant radius, they reflect off it while inside you, I'm thinking about making them begin with 0 radius and expand as they move away from you (which would also fix the fact that they appear in front of you at generation, terrible).
Propellant doesn't all fire in the same direction, there's a small amount of randomness to direction and magnitude (put through an atan function, so it still has a wide range but mostly deviates little), I like it. Currently, if you click and hold upwards, you can reach the moon (or Earth, if you'd like, it's blue because it's the second non-cueball colour in the list and yellow is the first) by clicking and holding upwards, try burning retrograde from it so the propellant reflects, it's quite beautiful.
The player also has an eye indicating the direction you're attempting to accelerate in while burning RCS, without the randomness (which affects you as well as it) or rounding to the nearest of eight directions (meant to operate from the mode 4 acceleration-based reference frame, but currently seems to be the absolute one). This will be meaningful when there will also be enemies that chase you (I'm going to use the per-frame missile interception equations to eliminate the component of velocity perpendicular to the direction from them to the target then maximise the component parallel, imagine the propellant is firing a very relatively massive missile each frame, if they have sufficient propulsion they should counteract gravity).
Also, in mode 4, the reference frame position is now properly that of the player, you no longer elongate (though when moving quickly you can flicker back to elongation, cause currently unknown). It's confusing when attempting to land on the moon, from my testing (after pushing it away from the planet), because you can't see that it's moving perpendicularly from the direction to you because it remains beneath you, it appears to fall away beneath you in a parabola, I think. This could be fixed by making reference frame angle have inertia, not yet sure if it should be done by taking a weighted average of the previously used and currently measured angle to get currently used angle or of the previously used and currently measured x and y accelerations to get the currently-used one (meaning switching from a weak position in one sphere of influence to a strong one will change direction more quickly and vice versa).
Also fix Pi calculator mode (previous change made it freeze to wait for a turn) and make n-body orbital velocity calculation use the sum of the mass of the parent and child, I'm going to make a child adjust the parent's parent planet's velocity as though it were a part of the child, it should work for simple solar systems where it remains reasonably circular and not 3-body-problem (binary star systems will be okay, but they seem to become less stable as you add more to the ring).