Fix collisions (now working), render circles instead of squares, make wall collisions work in sub-frames,
Collisions were broken because I forgot that the positional differences in the dot product weren't normalised (but they were in the resultant velocity setting), now they are. The division by the absolute value is done in the two different functions instead of squaring it because this way the programming is more readable and other functions can use the dot product part independently in the future without encountering the same problem.
Also, there can be arbitrarily many circles proceeding through colour wheel now (except brightness in each channel is cos(ball*360/balls-(channel*120)), which isn't accurate because the channel values are logarithmic (which is why correct colour summations are square-roots of sums of squares)).