Add quaternion functions for 3D (from Perspective3Dengine but now in Python)
quaternionMultiply uses the multiplication matrix typed out manually (I will make it iterate through its columns and rows in the future for additional elegance), quaternionConjugate negates the imaginary components, rotateByLocalGimbals will be for converting screen/keyboard drags (and enemy rotations (when your position is converted to a screen one in their cameras)) to quaternion acceleration. projectSpaceToScreen has four projections, the initial strange one, the azimuthal equidistant and equi-area and the stereographic (the latter of which has quite comedic variable names).
Also remove redundant range and indice in the line with the gravity coefficient calculation, and change `not (t<0 or t>timeToProceed)` to `0<=t<=timeToProceed` (which I didn't know about until now).